Fix up the front
We switched the rendering of the emails, but this has broken the
rendering of the front.

In the meantime, the front has shared a base with the admin page, but in
reality it should have it's own base.  This change seperates out the two
bases, and makes it clearer what's involved in rendering the front