Sort out authentication
AppEngine had a built in authentication model.
CloudRun doesn't come with that, so we need to create it.
We use Flask-Login which handles the OIDC stuff mostly, all we need is to add login_required for each of the views that login is required.
We then need the OAuth endpoints for handling login and callback.

Finally, we store the user details in the database because when a user returns, we don't want to send them back to Google just to get their UserID.  We probably could skip this, but it's interesting anyway.