Introduce the repositories and service model
Refactor the code between services and repositories, so mostly, data retrieval and creation is in the repo, and business rules are in the service.

We've done only the links here.  Many handlers still need to understand and handle a link object, and talk to the repo direct, which we'll clean up later if it makes sense